Array len functionalizer is carried into subpass arrayFunctions
