+++ tests/behaviour/expectations/index.ts
@@ -45,4 +45,4 @@ function filterTests(
   return tests.filter((test) => test.name.includes(filter));
 }
 
-export const expectations = filterTests(behaviour, semantic);
+export const expectations = filterTests(behaviour, semantic, 'solidity');
